refactor: header
Given that the bottom of the header (the dropdown) uses scripts, only the top part of the header is loaded.
The bottom part is loaded in the iframe. The WP site will both hide the top part of the header, and add padding to the bottom to keep it's position the same